backend/requirements.txt
Flask==3.0.3
flask-cors==4.0.1
gunicorn==22.0.0
python-dotenv==1.0.1
inference-sdk==0.9.17
requests==2.32.3
pillow==11.1.0
ultralytics==8.3.56